requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.0
django-ckeditor==6.7.2
django-js-asset==2.2.0
djangocms-admin-style==3.3.1
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
